Pull the annotation related code out into its own module
This commit only moves the code around.  There aren't any changes in the
moved code.

This better breaks up the bloated "internal_rep.rs" file.
